refactor: extract sentence builder and helpers
