+++ tests/test_back_step.py
@@ -1,22 +1,26 @@
 import ast
 from types import SimpleNamespace
 
+from src.draft_management import _draft_state_keys
 
-def _load_functions():
-    with open('a1sprechen.py', 'r', encoding='utf-8') as f:
+
+def _load_back_step():
+    with open("a1sprechen.py", "r", encoding="utf-8") as f:
         src = f.read()
     mod = ast.parse(src)
-    funcs = [node for node in mod.body if isinstance(node, ast.FunctionDef) and node.name in {'_draft_state_keys', 'back_step'}]
+    funcs = [
+        node for node in mod.body if isinstance(node, ast.FunctionDef) and node.name == "back_step"
+    ]
     module_ast = ast.Module(body=funcs, type_ignores=[])
-    code = compile(module_ast, 'a1sprechen.py', 'exec')
+    code = compile(module_ast, "a1sprechen.py", "exec")
     st = SimpleNamespace(session_state={})
-    glb = {'st': st}
+    glb = {"st": st, "_draft_state_keys": _draft_state_keys}
     exec(code, glb)
-    return glb['back_step'], glb['_draft_state_keys'], st
+    return glb["back_step"], st
 
 
 def test_back_step_clears_chat_state():
-    back_step, _draft_state_keys, st = _load_functions()
+    back_step, st = _load_back_step()
     ss = st.session_state
     ss.update({
         'falowen_stage': 4,
